requirements.txt
requests
python-dotenv
groq
google-search-results
schedule